+++ django_mysql/models/fields/__init__.py
@@ -1,3 +1,4 @@
+from django_mysql.models.fields.bit import *  # NOQA
 from django_mysql.models.fields.lists import *  # NOQA
 from django_mysql.models.fields.sets import *  # NOQA
 from django_mysql.models.fields.sizes import *  # NOQA